requirements.txt
flask
google.generativeai
deep_translator
gtts
pdfplumber